I have branched this source to contribute my VS2008 projects as a way of saying thanks to chmouel and the guys/gals responsible for this really handy library.
This has saved me a lot of time and I hope by contributing the VS2008 projects as an example to a fellow .NET noobie will do the same for them.

I came across this library via the following forum post:

http://stackoverflow.com/questions/4446956/how-to-upload-files-to-rackspace-cloud-using-windows-services

The example "upload-to-cf-cs" came from:

https://github.com/chmouel/upload-to-cf-cs

The cloud files library came from:

https://github.com/rackspace/csharp-cloudfiles

All credit to chmouel and jmeridth.